Add commands for adding and removing tasks and showing list
add-task requires the user to supply the task name and the due date,
then the task is added to the running list.

remove-task finds the task in the list and removes in from the list. the
user is asked to type the name of the task that they would like removed
and it is removed from the running list.

show-list dislays the list of all tasks.